requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
Flask==3.0.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mutagen==1.47.0
pycparser==2.21
pycryptodomex==3.20.0
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.2.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
yt-dlp==2023.12.30
